+++ src/parsing/request_header_parsing.cpp
@@ -6,7 +6,7 @@
 /*   By: jvacaris <jvacaris@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/06 20:31:03 by jvacaris          #+#    #+#             */
-/*   Updated: 2023/05/10 15:45:47 by jvacaris         ###   ########.fr       */
+/*   Updated: 2023/05/20 20:50:16 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@
 !			useless for now (except for debugging purposes). Most of it has been commented.
 !			If there's time, we can pass some explanation about the error to the returned page.
 */
-bool	Request::check_first_line_validity(std::string firstline)
+bool	Request::check_first_line_validity(std::string firstline, const ft::serverconf &_config)
 {
 	std::vector <std::string> params = cpp_split(firstline, ' ');
 
@@ -89,13 +89,45 @@ bool	Request::check_first_line_validity(std::string firstline)
 		// 	std::cerr << params[i] << " ";
 		// }
 		// std::cerr << TXT_RESET << std::endl;
+		method = -2;
 		return (false);
 	}
 	path.unparsed = params[1];
-	path.relative = correct_path(params[1]);
+	
+	
+	std::string path_iter;
+	path_iter = correct_path(params[1]);
+	std::vector<std::string> split_path = cpp_split(path_iter, '/');
+	unsigned int level = split_path.size();
+	while (path_iter.size() >= 0)	//!		Might enter an infinite loop. Check it out later.
+	{
+		it_location = _config.locations.find(path_iter);
+		if (it_location == _config.locations.end())
+		{
+			size_t last_slash = path_iter.find_last_of('/');
+			if (last_slash == std::string::npos)
+				path_iter = "";
+			else
+				path_iter = path_iter.substr(0, last_slash); //! Add a -1 after last_slash if is appends a '/'.
+			level--;
+		}
+		else
+			break;	
+	}
+
+	path.relative = correct_path(it_location->second.file_root);
+
+	while (level < split_path.size())
+	{
+		path.relative.append("/");
+		path.relative.append(split_path[level++]);
+	}
+	path.vec_relative = cpp_split(path.relative, '/');
 	path.absolute = SERVER_ROOT;
 	path.absolute.append("/");
 	path.absolute.append(path.relative);
+
+
 	return (true);
 }
 
@@ -127,10 +159,10 @@ bool check_header_validity(std::vector <std::string> line_vector, std::map<std::
 /*
 ?	Checks if all fo the testing worked out well.
 */
-bool	Request::check_request_validity(std::string fullheader)
+bool	Request::check_request_validity(std::string fullheader, const ft::serverconf &_config)
 {
 	std::vector <std::string> line_vector = cpp_split(fullheader, '\n');
-	if (!check_first_line_validity(line_vector[0]))
+	if (!check_first_line_validity(line_vector[0], _config))
 		return (false);
 	if (!check_header_validity(line_vector, header_map))
 		return (false);
@@ -143,12 +175,13 @@ bool	Request::check_request_validity(std::string fullheader)
 ?	the header and the body into a string.
 !	If there's no body, it will be left as an empty string ("").
 */
-void Request::header_parser()
+void Request::header_parser(const ft::serverconf &_config)
 {
 	size_t head_body_separation = fullrequest.find("\n\n");
-	if (!check_request_validity(fullrequest.substr(0, head_body_separation)))
+	if (!check_request_validity(fullrequest.substr(0, head_body_separation), _config))
 	{
-		method = -1;
+		if (method != -2)
+			method = -1;
 		return ;
 	}
 	if (head_body_separation != std::string::npos)