+++ src/requests/Request.cpp
@@ -6,7 +6,7 @@
 /*   By: jvacaris <jvacaris@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/30 20:13:54 by jvacaris          #+#    #+#             */
-/*   Updated: 2023/05/10 15:47:33 by jvacaris         ###   ########.fr       */
+/*   Updated: 2023/05/22 20:42:50 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,19 +21,58 @@
 	body = "";
 }*/
 
-Request::Request(const std::string &_input, const ft::serverconf &_config): fullrequest(_input), config(_config)
+Request::Request(const std::string &_input, const ft::serverconf &_config): config(_config), fullrequest(_input)
 {
-	header_parser();
+	header_parser(_config);
 }
 
-Request::Request(Request &tocopy): fullrequest(tocopy.fullrequest), config(tocopy.config), \
-method(tocopy.method), path(tocopy.path), header_map(tocopy.header_map), body(tocopy.body)
+Request::Request(Request &tocopy): config(tocopy.config), fullrequest(tocopy.fullrequest), \
+method(tocopy.method), path(tocopy.path), header_map(tocopy.header_map), body(tocopy.body), it_location(tocopy.it_location)
 {
 }
 
-Request::Request(const Request &tocopy): fullrequest(tocopy.fullrequest), config(tocopy.config), \
-method(tocopy.method), path(tocopy.path), header_map(tocopy.header_map), body(tocopy.body)
+Request::Request(const Request &tocopy): config(tocopy.config), fullrequest(tocopy.fullrequest), \
+method(tocopy.method), path(tocopy.path), header_map(tocopy.header_map), body(tocopy.body), it_location(tocopy.it_location)
 {
+}
+
+Request::~Request()
+{
+	
+}
+
+/*
+TODO		Hasn't been tested yet. 
+*/
+void Request::set_redirect_path()
+{
+	std::string try_path;
+	std::vector<std::string>::iterator ending = path.vec_relative.end();
+	
+	
+	while (true)
+	{
+		try_path = "";
+		for (std::vector<std::string>::iterator it = path.vec_relative.begin(); it != ending; it++)
+		{
+			if (it != path.vec_relative.begin())
+				try_path.append("/");
+			try_path.append(*it);
+		}
+		if (config.locations.find(try_path) != config.locations.end())
+		{
+			break ;
+		}
+		else
+		{
+			if (ending == path.vec_relative.begin())
+				return ;
+			else
+				ending--;
+		}
+	}
+	path.absolute = path.absolute.replace(path.absolute.find(try_path), try_path.length(), config.locations.find(try_path)->second.file_root);
+	path.relative = path.absolute;
 }
 
 const std::string &Request::get_fullrequest() const
@@ -82,5 +121,10 @@ const std::map<std::string, std::string> Request::get_headermap() const
 const std::string &Request::get_body() const
 {
 	return(body);
+}
+
+const std::map<std::string, ft::location>::const_iterator &Request::get_dir_params() const
+{
+	return (it_location);
 }