+++ src/responses/Response.hpp
@@ -6,7 +6,7 @@
 /*   By: jvacaris <jvacaris@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/01 20:44:32 by jvacaris          #+#    #+#             */
-/*   Updated: 2023/05/10 15:19:19 by jvacaris         ###   ########.fr       */
+/*   Updated: 2023/05/21 19:14:46 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,9 +32,8 @@ class Response
 		Response();
 
 	public:
-		Response(const Request _request);
+		Response(const Request &_request);
 		void return_content();
-		std::string reply();
 		std::string generate_response();
 		void file_status_custom_error(int file_status);
 		void return_error_message(int error_code);