+++ src/utils.cpp
@@ -6,7 +6,7 @@
 /*   By: emadriga <emadriga@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/06 20:30:58 by jvacaris          #+#    #+#             */
-/*   Updated: 2023/05/10 16:08:37 by emadriga         ###   ########.fr       */
+/*   Updated: 2023/05/23 15:31:37 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -356,7 +356,6 @@ std::string correct_path(const std::string &orig_path)
 				curr_route--;
 				routes.erase(curr_route);
 			}
-
 		}
 		else if (*curr_route == ".")
 			routes.erase(curr_route);
@@ -376,7 +375,6 @@ std::string correct_path(const std::string &orig_path)
 	if (retval[0] == '/')
 		retval = retval.substr(1, retval.size() - 1);
 	return (retval);
-
 }
 
 //! Debugging purposes, please delete before evaluating.